assignment3/requirements.txt
langchain>=0.3.7
langchain-community>=0.3.7
langgraph>=0.2.45
langchain-core>=0.3.15
langchain-chroma>=0.1.4
langchain-google-genai>=2.0.4
langchain-openai>=0.2.6
langchain-huggingface>=0.1.0
langchain-anthropic>=0.3.0
transformers>=4.46.1
sentence-transformers
chromadb
tenacity
termcolor
pypdf
pymupdf
python-dotenv
